+++ src/scenes/Home/index.js
@@ -1,5 +1,9 @@
 import React from "react";
 
+import { ArtistCard } from "components/ArtistCard";
+
+import "./styles.css";
+
 export default class Home extends React.Component {
   constructor() {
     super();
@@ -8,14 +12,71 @@ export default class Home extends React.Component {
 
   componentDidMount() {
     // Change page title
-    document.title = "New New"
+    document.title = "New New";
   }
 
   render() {
     return (
       <div className="container">
-        <h1>Home</h1>
+        <div className="flex-center">
+          <div className="row">
+            <div className="col-sm-4 artist-card">
+              <ArtistCard
+                name="Atu"
+                image="https://i.imgur.com/9z6f0IY.png"
+                noteableTitle="Sonder Unreleased 01"
+                noteableLink="https://www.youtube.com/watch?v=t6QanTjf0s0"
+                submitter="@bryanbrotonel"
+              />
+            </div>
+            <div className="col-sm-4 artist-card">
+              <ArtistCard
+                name="Dpat"
+                image="https://i.imgur.com/CM278ie.jpg"
+                noteableTitle="Sonder Unreleased 05-06"
+                noteableLink="https://www.youtube.com/watch?v=RXfF-bQxPp4"
+                submitter="@bryanbrotonel"
+              />
+            </div>
+            <div className="col-sm-4 artist-card">
+              <ArtistCard
+                name="Brent Faiyaz"
+                image="https://i.imgur.com/HERCJtO.jpg"
+                noteableTitle="Poison"
+                noteableLink="https://soundcloud.com/brentfaiyaz/poison-prod-by-ben-free?in=brentfaiyaz/sets/a-m-paradox"
+                submitter="@bryanbrotonel"
+              />
+            </div>
+            <div className="col-sm-4 artist-card">
+              <ArtistCard
+                name="Blood Orange"
+                image="https://i.imgur.com/Ao40HHG.jpg"
+                noteableTitle="Saint"
+                noteableLink="https://soundcloud.com/bloodorange/saint?in=bloodorange/sets/negro-swan"
+                submitter="@bryanbrotonel"
+              />
+            </div>
+            <div className="col-sm-4 artist-card">
+              <ArtistCard
+                name="Arin Ray"
+                image="https://i.imgur.com/A6npeZC.jpg"
+                noteableTitle="Stressin"
+                noteableLink="https://soundcloud.com/arin-ray/stressin?in=arin-ray/sets/platinum-fire-1"
+                submitter="@bryanbrotonel"
+              />
+            </div>
+            <div className="col-sm-4 artist-card">
+              <ArtistCard
+                name="Rex Orange County"
+                image="https://i.imgur.com/KCxr908.jpg"
+                noteableTitle="Edition"
+                noteableLink="https://www.youtube.com/watch?v=8ECLpFVE5pE"
+                submitter="@bryanbrotonel"
+              />
+            </div>
+          </div>
+        </div>
       </div>
-    )
+    );
   }
 }