+++ src/scenes/Share/index.js
@@ -29,14 +29,11 @@ export default class Share extends React.Component {
                 </h5>
               </div>
               <div>
-                <p>
-                  <h6>Sharing music is for the soul.</h6>
-                </p>
+                <h6>Sharing music is for the soul.</h6>
                 <p>
                   From one listener to another, let's share music. Drop your
-                  current favourite artsit and your favourite work from them,
-                  and I'll add it to the New New collection. Also, don't forget
-                  to give me your name so I can shout you out.
+                  current favourite artist and your favourite work from them,
+                  and I'll add it to the New New collection. Also, don't forget to drop your handle so I can shout you out.
                 </p>
                 <p>Love.</p>
               </div>