+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.0.3 - 2018-08-26
 ### Added
 - Added motive page layout
+- New New logo and favicon
 
 ## 0.0.2 - 2018-08-21, 2018-08-26
 ### Added