+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.0.4 - 2018-08-27
+### Added
+- Added home page layout
+- Added artist card component
+
+### Changed
+- Changed sharing name to handle
+
 ## 0.0.3 - 2018-08-26
 ### Added
 - Added motive page layout